Refactor tests to use common data/fixtures
Refactor to reduce copying similar data used acrossed different
tests.

Ticket: AT-7457
