+++ api/cost/start-cost-job.test.ts
@@ -4,32 +4,13 @@ import { ApiSuccessResponse, SuccessStatusCode, ValidationErrorResponse } from "
 import { handler } from "./start-cost-job";
 import { sqsClient } from "../../utils/aws-sdk/sqs";
 import { SendMessageCommand } from "@aws-sdk/client-sqs";
-import { CostRequest } from "../../models/cost-jobs";
 import { Network } from "../../models/cloud-service-providers";
+import { validCostRequest, baseApiRequest } from "../util/common-test-fixtures";
 
 const sqsMock = mockClient(sqsClient);
 beforeEach(() => {
   sqsMock.reset();
 });
-
-const validCostRequest: CostRequest = {
-  requestId: "81b31a89-e3e5-46ee-acfe-75436bd14577",
-  portfolioId: "b02e77d1-234d-4e3d-bc85-b57ca5a93952",
-  targetCsp: {
-    name: "CSP_A",
-    uri: "http://www.somecspvendor.com/api/atat",
-    network: Network.NETWORK_1,
-  },
-  startDate: "2022-01-01",
-  endDate: "2022-12-01",
-};
-const requestContext = { identity: { sourceIp: "203.0.113.0" } };
-const baseApiRequest = {
-  headers: {
-    "Content-Type": "application/json",
-  },
-  requestContext,
-} as any;
 
 describe("Cost request operations", () => {
   it("should return successfully if no error enqueueing message", async () => {
