+++ api/provision/start-provisioning-job.test.ts
@@ -1,52 +1,18 @@
 import { Context } from "aws-lambda";
 import { mockClient } from "aws-sdk-client-mock";
-import { Network } from "../../models/cloud-service-providers";
 import { ProvisionRequestType } from "../../models/provisioning-jobs";
 import { sfnClient } from "../../utils/aws-sdk/step-functions";
 import { ApiSuccessResponse, ValidationErrorResponse } from "../../utils/response";
+import {
+  fundingSources,
+  provisioningBodyNoPayload,
+  validRequest,
+  // requestContext,
+  operators,
+} from "../util/common-test-fixtures";
 import { handler } from "./start-provisioning-job";
 
-export const fundingSources = [
-  {
-    taskOrderNumber: "1234567890123",
-    clin: "9999",
-    popStartDate: "2021-07-01",
-    popEndDate: "2022-07-01",
-  },
-];
-export const operators = ["admin1@mail.mil", "superAdmin@mail.mil"];
-export const provisioningBodyNoPayload = {
-  jobId: "81b31a89-e3e5-46ee-acfe-75436bd14577",
-  userId: "21d18790-bf3e-4529-a361-460ee6d16e0b",
-  portfolioId: "b02e77d1-234d-4e3d-bc85-b57ca5a93952",
-  operationType: ProvisionRequestType.ADD_OPERATORS,
-  targetCsp: {
-    name: "CSP_A",
-    uri: "http://www.somecspvendor.com/api/atat",
-    network: Network.NETWORK_1,
-  },
-  cspInvocation: undefined,
-  cspResponse: undefined,
-};
-
-export const provisioningBodyWithPayload = {
-  ...provisioningBodyNoPayload,
-  payload: {
-    name: "Sample Portfolio",
-    fundingSources,
-    operators,
-  },
-};
 export const requestContext = { identity: { sourceIp: "203.0.113.0" } };
-
-export const validRequest = {
-  body: JSON.stringify(provisioningBodyWithPayload),
-  headers: {
-    "Content-Type": "application/json",
-  },
-  requestContext,
-} as any;
-
 const sfnMock = mockClient(sfnClient);
 beforeEach(() => {
   sfnMock.reset();
