+++ api/util/csp-request.test.ts
@@ -4,10 +4,10 @@ import * as cspConfig from "../provision/csp-configuration";
 import { Network } from "../../models/cloud-service-providers";
 import { CostRequest, CspRequest, CspRequestType } from "../../models/cost-jobs";
 import { ErrorStatusCode, SuccessStatusCode } from "../../utils/response";
-import { validRequestBody } from "../cost/cost-request-fn.test";
 import { constructProvisionRequestForCsp } from "../provision/csp-write-portfolio.test";
 import { cspRequest } from "./csp-request";
 import { ProvisionRequest } from "../../models/provisioning-jobs";
+import { validCostRequest } from "../util/common-test-fixtures";
 
 // Reused mocks
 jest.mock("../provision/csp-configuration");
@@ -21,7 +21,7 @@ describe("Successful invocation of mock CSP function", () => {
   const fakeGoodResponse = { fake: "csp response" };
   it("should return 200 when COST csp request", async () => {
     // GIVEN
-    const request: CspRequestType<CostRequest> = { requestType: CspRequest.COST, body: validRequestBody };
+    const request: CspRequestType<CostRequest> = { requestType: CspRequest.COST, body: validCostRequest };
     mockedConfig.mockResolvedValueOnce({ uri: "https://mockcsp.cspa/atat/" });
     mockedAxios.post.mockResolvedValueOnce({
       data: fakeGoodResponse,
@@ -74,7 +74,7 @@ describe("Failed CSP invocation operations", () => {
       const request = {
         requestType,
         body: {
-          ...validRequestBody,
+          ...validCostRequest,
           targetCsp: { name: "CSP_BAD", uri: "https://csp.com", network: Network.NETWORK_1 },
         },
       } as any;
@@ -96,7 +96,7 @@ describe("Failed CSP invocation operations", () => {
     async (requestType) => {
       // GIVEN
       const fakeBadResponse = { bad: "fake csp response" };
-      const request = { requestType, body: validRequestBody };
+      const request = { requestType, body: validCostRequest };
       mockedGetToken.mockResolvedValueOnce({ access_token: "FakeToken", expires_in: 0, token_type: "Bearer" });
       mockedConfig.mockResolvedValueOnce({ uri: "https://mockcsp.cspa/atat/" });
       mockedAxios.post.mockResolvedValueOnce({
@@ -119,7 +119,7 @@ describe("Failed CSP invocation operations", () => {
     "should throw a 500 error when a CSP internal error occurs",
     async (requestType) => {
       const internalErrorResponse = { some: "internal error" };
-      const request = { requestType, body: validRequestBody };
+      const request = { requestType, body: validCostRequest };
       mockedGetToken.mockResolvedValueOnce({ access_token: "FakeToken", expires_in: 0, token_type: "Bearer" });
       mockedConfig.mockResolvedValueOnce({ uri: "https://mockcsp.cspa/atat/" });
       mockedAxios.post.mockResolvedValueOnce({
